examples/object_detection/requirements.txt
pytorch-ignite
torchvision
fire
pycocotools
albumentations
tensorboard
clearml
ultralytics